+++ tests/test_calculator.py
@@ -1,8 +1,8 @@
-from calculator import add, div, mul, sub
+from src.calculator import add, div, mul, sub
 
 
 def test_add():
-    assert add(1, 1) == 2
+    assert add(1, 1) == 1
 
 
 def test_sub():
